requirements.txt
SpeechRecognition 
pyttsx3 
groq
PyAudio
python-dotenv
keyboard
psutil 
requests
wikipedia
IMDbPY
bs4
httpx
langdetect
deep-translator
googletrans==3.1.0a0
schedule 
playsound
pygame